+++ libs/email.js
@@ -31,11 +31,24 @@ module.exports = function (app) {
     var templateRootLocal = app.get('EMAIL_TEMPLATE_ROOT_LOCAL');
 
     var emailHeaderLogoName = 'logo.png';
+    var emailFooterLogoName = 'logo_footer.png';
     var emailHeaderLogo = path.join(templateRoot, 'images/logo-email.png');
+    var emailFooterLogo = path.join(templateRoot, 'images/logo-email-small.png');
     if (fs.existsSync(path.join(templateRootLocal, 'images/logo-email.png'))) { //eslint-disable-line no-sync
         emailHeaderLogo = path.join(templateRootLocal, 'images/logo-email.png');
     }
     logger.info('Using email header logo from', emailHeaderLogo);
+
+    var linkToPlatform = config.email.linkToPlatform || urlLib.getFe();
+    var linkToPrivacyPolicy = config.email.linkToPrivacyPolicy;
+    var linkedData = {
+        footerLinks: {
+            linkToPlatform: linkToPlatform,
+            linkToPrivacyPolicy: linkToPrivacyPolicy
+        }
+    };
+    var styles = config.email.styles;
+    var partnerStyles = config.email.partnerStyles;
 
     var templateCache = {};
 
@@ -108,7 +121,7 @@ module.exports = function (app) {
 
                 _.forEach(users, function (user) {
                     var templateObject = resolveTemplate('accountVerification', user.language);
-
+                    linkedData.translations = templateObject.translations;
                     var linkVerify = urlLib.getApi('/api/auth/verify/:code', {code: emailVerificationCode}, {token: token});
 
                     // https://github.com/bevacqua/campaign#email-sending-option
@@ -120,15 +133,23 @@ module.exports = function (app) {
                             {
                                 name: emailHeaderLogoName,
                                 file: emailHeaderLogo
+                            },
+                            {
+                                name: emailFooterLogoName,
+                                file: emailFooterLogo
                             }
                         ],
                         //Placeholders
                         toUser: user,
                         linkVerify: linkVerify,
                         linkToApplication: urlLib.getFe(),
+                        linkToPlatform: linkToPlatform,
+                        linkToPrivacyPolicy: linkToPrivacyPolicy,
                         provider: {
                             merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                        }
+                        },
+                        styles: styles,
+                        linkedData: linkedData
                     });
 
                     promisesToResolve.push(userEmailPromise);
@@ -160,7 +181,7 @@ module.exports = function (app) {
 
                 _.forEach(users, function (user) {
                     var template = resolveTemplate('passwordReset', user.language);
-
+                    linkedData.translations = template.translations;
                     var userEmailPromise = emailClient.sendStringAsync(template.body, {
                         subject: template.translations.PASSWORD_RESET.SUBJECT,
                         to: user.email,
@@ -168,6 +189,10 @@ module.exports = function (app) {
                             {
                                 name: emailHeaderLogoName,
                                 file: emailHeaderLogo
+                            },
+                            {
+                                name: emailFooterLogoName,
+                                file: emailFooterLogo
                             }
                         ],
                         social: config.email.social, // social.name maps to "from_name". I think this should be part of Campaign client config OR at least defaults can be set somewhere in client options
@@ -175,9 +200,13 @@ module.exports = function (app) {
                         toUser: user,
                         linkReset: urlLib.getFe('/account/password/reset/:passwordResetCode', {passwordResetCode: passwordResetCode}, {email: user.email}),
                         linkToApplication: urlLib.getFe(),
+                        linkToPlatform: linkToPlatform,
+                        linkToPrivacyPolicy: linkToPrivacyPolicy,
                         provider: {
                             merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                        }
+                        },
+                        styles: styles,
+                        linkedData: linkedData
                     });
 
                     promisesToResolve.push(userEmailPromise);
@@ -213,7 +242,7 @@ module.exports = function (app) {
 
             return Promise.resolve();
         }
-
+        var customStyles = styles;
         var toUsersPromise = User.findAll({
             where: {
                 id: toUserIds
@@ -270,6 +299,7 @@ module.exports = function (app) {
                         logoFile = templateRoot + '/images/logo-email_' + sourceSite + '.png';
                         templateName = 'inviteTopic_' + sourceSite;
                         linkToApplication = partner.website;
+                        customStyles = partnerStyles[sourceSite];
                     }
 
                     //TODO: we can win performance if we collect together all Users with same language and send these with 1 request to mail provider
@@ -306,6 +336,8 @@ module.exports = function (app) {
                             // In case Topic has no title, just show the full url.
                             topic.title = topic.title ? topic.title : linkViewTopic;
 
+                            linkedData.translations = template.translations;
+
                             var emailPromise = emailClient.sendStringAsync(template.body, {
                                 from: from,
                                 subject: subject,
@@ -314,6 +346,10 @@ module.exports = function (app) {
                                     {
                                         name: emailHeaderLogoName,
                                         file: logoFile
+                                    },
+                                    {
+                                        name: emailFooterLogoName,
+                                        file: emailFooterLogo
                                     }
                                 ],
                                 social: config.email.social, // social.name maps to "from_name". I think this should be part of Campaign client config OR at least defaults can be set somewhere in client options
@@ -322,9 +358,13 @@ module.exports = function (app) {
                                 topic: topic,
                                 linkViewTopic: linkViewTopic,
                                 linkToApplication: linkToApplication,
+                                linkToPlatform: linkToPlatform,
+                                linkToPrivacyPolicy: linkToPrivacyPolicy,
                                 provider: {
                                     merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                                }
+                                },
+                                styles: customStyles,
+                                linkedData: linkedData
                             });
 
                             promisesToResolve.push(emailPromise);
@@ -414,6 +454,7 @@ module.exports = function (app) {
                     _.forEach(toUsers, function (user) {
                         if (user.email) {
                             var template = resolveTemplate('inviteTopic', user.language);
+                            linkedData.translations = template.translations;
                             // TODO: Could use Mu here....
                             var subject = template.translations.INVITE_TOPIC.SUBJECT
                                 .replace('{{fromUser.name}}', util.escapeHtml(fromUser.name));
@@ -433,6 +474,10 @@ module.exports = function (app) {
                                     {
                                         name: emailHeaderLogoName,
                                         file: emailHeaderLogo
+                                    },
+                                    {
+                                        name: emailFooterLogoName,
+                                        file: emailFooterLogo
                                     }
                                 ],
                                 social: config.email.social, // social.name maps to "from_name". I think this should be part of Campaign client config OR at least defaults can be set somewhere in client options
@@ -441,9 +486,13 @@ module.exports = function (app) {
                                 topic: topic,
                                 linkViewTopic: linkViewTopic,
                                 linkToApplication: urlLib.getFe(),
+                                linkToPlatform: linkToPlatform,
+                                linkToPrivacyPolicy: linkToPrivacyPolicy,
                                 provider: {
                                     merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                                }
+                                },
+                                styles: styles,
+                                linkedData: linkedData
                             });
 
                             promisesToResolve.push(sendEmailPromise);
@@ -517,6 +566,7 @@ module.exports = function (app) {
                     _.forEach(toUsers, function (user) {
                         if (user.email) {
                             var template = resolveTemplate('inviteGroup', user.language);
+                            linkedData.translations = template.translations;
                             // TODO: could use Mu here...
                             var subject = template.translations.INVITE_GROUP.SUBJECT
                                 .replace('{{fromUser.name}}', util.escapeHtml(fromUser.name))
@@ -529,6 +579,10 @@ module.exports = function (app) {
                                     {
                                         name: emailHeaderLogoName,
                                         file: emailHeaderLogo
+                                    },
+                                    {
+                                        name: emailFooterLogoName,
+                                        file: emailFooterLogo
                                     }
                                 ],
                                 social: config.email.social, // social.name maps to "from_name". I think this should be part of Campaign client config OR at least defaults can be set somewhere in client options
@@ -543,7 +597,9 @@ module.exports = function (app) {
                                 linkToApplication: urlLib.getFe(),
                                 provider: {
                                     merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                                }
+                                },
+                                styles: styles,
+                                linkedData: linkedData
                             });
 
                             promisesToResolve.push(userEmailPromise);
@@ -700,7 +756,7 @@ module.exports = function (app) {
                 var commentCreatorInformed = true;
                 if (commentInfo.comment.creator.email) {
                     var templateObject = resolveTemplate('reportCommentCreator', commentInfo.comment.creator.language);
-
+                    linkedData.translations = templateObject.translations;
                     var linkViewTopic = urlLib.getFe('/topics/:topicId', {topicId: commentInfo.topic.id});
 
                     var promiseCreatorEmail = emailClient.sendStringAsync(
@@ -713,6 +769,10 @@ module.exports = function (app) {
                                 {
                                     name: emailHeaderLogoName,
                                     file: emailHeaderLogo
+                                },
+                                {
+                                    name: emailFooterLogoName,
+                                    file: emailFooterLogo
                                 }
                             ],
                             //Placeholders
@@ -725,7 +785,9 @@ module.exports = function (app) {
                             linkToApplication: urlLib.getFe(),
                             provider: {
                                 merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                            }
+                            },
+                            styles: styles,
+                            linkedData: linkedData
                         }
                     );
                     promisesToResolve.push(promiseCreatorEmail);
@@ -774,6 +836,10 @@ module.exports = function (app) {
                                         {
                                             name: emailHeaderLogoName,
                                             file: emailHeaderLogo
+                                        },
+                                        {
+                                            name: emailFooterLogoName,
+                                            file: emailFooterLogo
                                         }
                                     ],
                                     //Placeholders...
@@ -785,9 +851,12 @@ module.exports = function (app) {
                                     linkModerate: linkModerate + '?token=' + encodeURIComponent(token),
                                     isUserNotified: commentCreatorInformed,
                                     linkToApplication: urlLib.getFe(),
+                                    linkToPlatform: linkToPlatform,
+                                    linkToPrivacyPolicy: linkToPrivacyPolicy,
                                     provider: {
                                         merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                                    }
+                                    },
+                                    styles: styles
                                 }
                             );
 
@@ -821,6 +890,7 @@ module.exports = function (app) {
         }
 
         var template = resolveTemplate('toParliament', 'et'); // Estonian Gov only accepts et
+        linkedData.translations = template.translations;
         var linkToApplication = config.features.sendToParliament.urlPrefix;
 
         var from = config.features.sendToParliament.from;
@@ -830,6 +900,11 @@ module.exports = function (app) {
         var logoFile = templateRoot + '/images/logo-email_rahvaalgatus.ee.png';
 
         var promisesToResolve = [];
+        var customStyles = {
+            headerBackgroundColor: '#252525',
+            logoWidth: 360,
+            logoHeight: 51
+        };
 
         // Email to Parliament
         var emailToParliamentPromise = emailClient
@@ -843,6 +918,10 @@ module.exports = function (app) {
                         {
                             name: emailHeaderLogoName,
                             file: logoFile
+                        },
+                        {
+                            name: emailFooterLogoName,
+                            file: emailFooterLogo
                         }
                     ],
                     social: config.email.social, // social.name maps to "from_name". I think this should be part of Campaign client config OR at least defaults can be set somewhere in client options
@@ -852,11 +931,15 @@ module.exports = function (app) {
                     linkDownloadBdocFinalExpiryDate: moment(linkDownloadBdocFinalExpiryDate).locale('et').format('LL'),
                     linkAddEvent: linkAddEvent,
                     linkToApplication: linkToApplication,
+                    linkToPlatform: linkToPlatform,
+                    linkToPrivacyPolicy: linkToPrivacyPolicy,
                     topic: topic,
                     contact: contact,
                     provider: {
                         merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                    }
+                    },
+                    styles: customStyles,
+                    linkedData: linkedData
                 }
             )
             .then(function () {
@@ -882,20 +965,28 @@ module.exports = function (app) {
                         {
                             name: emailHeaderLogoName,
                             file: logoFile
+                        },
+                        {
+                            name: emailFooterLogoName,
+                            file: emailFooterLogo
                         }
                     ],
                     social: config.email.social, // social.name maps to "from_name". I think this should be part of Campaign client config OR at least defaults can be set somewhere in client options
                     //Placeholders..
                     linkViewTopic: linkViewTopic,
-                    linkDownloadBdocFinal: linkDownloadBdocFinal,
-                    linkDownloadBdocFinalExpiryDate: moment(linkDownloadBdocFinalExpiryDate).locale('et').format('LL'),
+                    linkDownloadBdocFinal: config.features.sendToParliament.sendContainerDownloadLinkToCreator ? linkDownloadBdocFinal : null,
+                    linkDownloadBdocFinalExpiryDate: config.features.sendToParliament.sendContainerDownloadLinkToCreator ? moment(linkDownloadBdocFinalExpiryDate).locale('et').format('LL') : null,
                     linkAddEvent: null,
                     linkToApplication: linkToApplication,
+                    linkToPlatform: linkToPlatform,
+                    linkToPrivacyPolicy: linkToPrivacyPolicy,
                     topic: topic,
                     contact: contact,
                     provider: {
                         merge: {} // TODO: empty merge required until fix - https://github.com/bevacqua/campaign-mailgun/issues/1
-                    }
+                    },
+                    styles: customStyles,
+                    linkedData: linkedData
                 }
             )
             .then(function () {